requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-cors==0.8.1
aiokafka==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==7.1.2
colorclass==2.2.2
colorlog==6.9.0
croniter==2.0.7
cryptography==46.0.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
faust-streaming==0.11.3
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
intervaltree==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==1.4.7
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mode==4.3.2
mode-streaming==0.4.1
multidict==6.6.4
mypy_extensions==1.1.0
opentracing==1.3.0
orjson==3.11.3
packaging==25.0
passlib==1.7.4
propcache==0.3.2
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
robinhood-aiokafka==1.1.6
rsa==4.9.1
sentry-sdk==2.38.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
starlette==0.47.3
terminaltables==3.1.10
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
venusian==3.1.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
